refactor: Move Solver module into Math module as a sub section
Math module is the home for more advance math operations.
For now, we have Eigen, Solvers, and RANSAC sub sections,
more sections (Statistics) are on the way

Others:
1. Remove FFTW3 dependencies from the project
FFTW3 is not straightforward to build from source, and link.
FFT from Eigen is used instead, it can use FFTW3 as backend anyway.